Indigo.Compilation

Description

This module contains everything related to compilation from Indigo to Lorentz, including plain Indigo code, as well as Indigo contracts.

Documentation

compileIndigo :: forall n inp out a. (SingI (ToPeano n), Default (MetaData inp), AreIndigoParams (ToPeano n) inp, TypeValue a) => IndigoWithParams (ToPeano n) inp out a -> inp :-> out

Compile Indigo code to Lorentz.

Note: it is necessary to specify the number of parameters (using the first type variable) of the Indigo function. Also, these should be on the top of the input stack in inverse order (see IndigoWithParams).

type IndigoContract param st = IndigoContract' param st '[param, st, Ops]

A utility type for a contract written in Indigo.

Note that is a more restrictive type than what is needed for compilation with compileIndigoContract and it's mainly intended to hide arbitrary outputs and be similar to Lorentz's ContractCode.

type IndigoContract' param st out = (HasStorage st, HasSideEffects) => Var param -> IndigoM '[param, st, Ops] out ()

Type of a contract that can be compiled to Lorentz with compileIndigoContract.

compileIndigoContract :: forall param st out. (TypeValue param, TypeValue st) => IndigoContract' param st out -> ContractCode param st

Compile Indigo code to Lorentz contract. Drop elements from the stack to return only [Operation] and Storage.

type Ops = [Operation]

type HasSideEffects = Given (Var Ops)

Allows to get a variable with operations

operationsVar :: HasSideEffects => Var Ops

Return a variable which refers to a stack cell with operations

type HasStorage st = Given (Var st)

Allows to get a variable with storage

storageVar :: HasStorage st => Var st

Return a variable which refers to a stack cell with storage