Indigo.Expr

Description

Expressions supported in Indigo language and their compilation to Lorentz code.

This module contains only basic building blocks that can be used to implement anything else. Other modules provide high level language constructions and standard functions.

Expr data type and custom expressions

data Expr a where

Constructors

C :: (NiceConstant a, Typeable a) => a -> Expr a

V :: TypeValue a => Var a -> Expr a

Cast :: (IsExpr ex a, KnownValue a, Typeable a) => ex -> Expr a

Add :: (AreExprs ex1 ex2 n m, ArithOpHs Add n m, TypeValue (ArithResHs Add n m)) => ex1 -> ex2 -> Expr (ArithResHs Add n m)

Sub :: (AreExprs ex1 ex2 n m, ArithOpHs Sub n m, TypeValue (ArithResHs Sub n m)) => ex1 -> ex2 -> Expr (ArithResHs Sub n m)

Mul :: (AreExprs ex1 ex2 n m, ArithOpHs Mul n m, TypeValue (ArithResHs Mul n m)) => ex1 -> ex2 -> Expr (ArithResHs Mul n m)

Div :: (AreExprs ex1 ex2 n m, EDivOpHs n m, TypeValue (EDivOpResHs n m)) => ex1 -> ex2 -> Expr (EDivOpResHs n m)

Mod :: (AreExprs ex1 ex2 n m, EDivOpHs n m, TypeValue (EModOpResHs n m)) => ex1 -> ex2 -> Expr (EModOpResHs n m)

Abs :: (IsExpr ex n, UnaryArithOpHs Abs n, TypeValue (UnaryArithResHs Abs n)) => ex -> Expr (UnaryArithResHs Abs n)

Neg :: (IsExpr ex n, UnaryArithOpHs Neg n, TypeValue (UnaryArithResHs Neg n)) => ex -> Expr (UnaryArithResHs Neg n)

Lsl :: (AreExprs ex1 ex2 n m, ArithOpHs Lsl n m, TypeValue (ArithResHs Lsl n m)) => ex1 -> ex2 -> Expr (ArithResHs Lsl n m)

Lsr :: (AreExprs ex1 ex2 n m, ArithOpHs Lsr n m, TypeValue (ArithResHs Lsr n m)) => ex1 -> ex2 -> Expr (ArithResHs Lsr n m)

Eq' :: (AreExprs ex1 ex2 n n, NiceComparable n) => ex1 -> ex2 -> Expr Bool

Neq :: (AreExprs ex1 ex2 n n, NiceComparable n) => ex1 -> ex2 -> Expr Bool

Le :: (AreExprs ex1 ex2 n n, NiceComparable n) => ex1 -> ex2 -> Expr Bool

Lt :: (AreExprs ex1 ex2 n n, NiceComparable n) => ex1 -> ex2 -> Expr Bool

Ge :: (AreExprs ex1 ex2 n n, NiceComparable n) => ex1 -> ex2 -> Expr Bool

Gt :: (AreExprs ex1 ex2 n n, NiceComparable n) => ex1 -> ex2 -> Expr Bool

Or :: (AreExprs ex1 ex2 n m, ArithOpHs Or n m, TypeValue (ArithResHs Or n m)) => ex1 -> ex2 -> Expr (ArithResHs Or n m)

Xor :: (AreExprs ex1 ex2 n m, ArithOpHs Xor n m, TypeValue (ArithResHs Xor n m)) => ex1 -> ex2 -> Expr (ArithResHs Xor n m)

And :: (AreExprs ex1 ex2 n m, ArithOpHs And n m, TypeValue (ArithResHs And n m)) => ex1 -> ex2 -> Expr (ArithResHs And n m)

Not :: (IsExpr op n, UnaryArithOpHs Not n, TypeValue (UnaryArithResHs Not n)) => op -> Expr (UnaryArithResHs Not n)

Int' :: IsExpr ex Natural => ex -> Expr Integer

IsNat :: IsExpr ex Integer => ex -> Expr (Maybe Natural)

Fst :: (IsExpr op (n, m), TypeValue n) => op -> Expr n

Snd :: (IsExpr op (n, m), TypeValue m) => op -> Expr m

Pair :: (AreExprs ex1 ex2 n m, TypeValue (n, m)) => ex1 -> ex2 -> Expr (n, m)

Some :: (IsExpr ex t, TypeValue (Maybe t)) => ex -> Expr (Maybe t)

None :: (KnownValue t, TypeValue t) => Expr (Maybe t)

Mem :: (MemOpHs c, IsExpr exc c, IsExpr exck (MemOpKeyHs c)) => exck -> exc -> Expr Bool

StToField :: (StoreHasField store fname ftype, TypeValue ftype, IsExpr exStore store) => Label fname -> exStore -> Expr ftype

ULookup :: (HasUStore name key value store, IsExpr exKey key, IsExpr exStore (UStore store), TypeValue value) => Label name -> exKey -> exStore -> Expr (Maybe value)

UInsertNew :: (HasUStore name key value store, IsError err, IsExpr exKey key, IsExpr exVal value, IsExpr exStore (UStore store), TypeValue (UStore store)) => Label name -> err -> exKey -> exVal -> exStore -> Expr (UStore store)

UInsert :: (HasUStore name key value store, IsExpr exKey key, IsExpr exVal value, IsExpr exStore (UStore store), TypeValue (UStore store)) => Label name -> exKey -> exVal -> exStore -> Expr (UStore store)

ToField :: (InstrGetFieldC dt name, IsExpr exDt dt, TypeValue (GetFieldType dt name)) => exDt -> Label name -> Expr (GetFieldType dt name)

SetField :: (InstrSetFieldC dt name, IsExpr exDt dt, IsExpr exFld (GetFieldType dt name), TypeValue dt) => exDt -> Label name -> exFld -> Expr dt

Name :: (IsExpr ex t, TypeValue (name :! t)) => Label name -> ex -> Expr (name :! t)

UnName :: (IsExpr ex (name :! t), TypeValue t) => Label name -> ex -> Expr t

Construct :: (InstrConstructC dt, RMap (ConstructorFieldTypes dt), TypeValue dt) => Rec Expr (ConstructorFieldTypes dt) -> Expr dt

ConstructT :: (InstrConstructC dt, TypeValue dt, RMap (ConstructorFieldTypes dt), fields ~ Rec Expr (ConstructorFieldTypes dt), RecFromTuple fields) => IsoRecTuple fields -> Expr dt
Construct a record from a tuple of named indigo parameters. For instance: ``` constructT @("sender" :: Address, "amount" :: Natural) ( (sender :: Expr Address) !~ #sender , (constExpr 10 :: Expr Natural) !~ #amount ) ```

SetContains :: (IsExpr exKey key, IsExpr exSet (Set key), IsComparable key) => exKey -> exSet -> Expr Bool

SetInsert :: (IsExpr exKey key, IsExpr exSet (Set key), IsComparable key, TypeValue (Set key)) => exKey -> exSet -> Expr (Set key)

SetRemove :: (IsExpr exKey key, IsExpr exSet (Set key), IsComparable key, TypeValue (Set key)) => exKey -> exSet -> Expr (Set key)

SetSize :: IsExpr exSet (Set key) => exSet -> Expr Natural

MapGet :: (GetOpHs c, TypeValue (Maybe (GetOpValHs c)), IsExpr exKey (GetOpKeyHs c), IsExpr exMap c) => exKey -> exMap -> Expr (Maybe (GetOpValHs c))

MapUpdate :: (UpdOpHs c, TypeValue c, IsExpr exKey (UpdOpKeyHs c), IsExpr exVal (UpdOpParamsHs c), IsExpr exMap c) => exKey -> exVal -> exMap -> Expr c

MapSize :: (SizeOpHs c, IsExpr exMap c) => exMap -> Expr Natural

BigMapGet :: (IsExpr exKey key, IsExpr exBigMap (BigMap key value), IsComparable key, TypeValue (Maybe value)) => exKey -> exBigMap -> Expr (Maybe value)

BigMapUpdate :: (IsExpr exKey key, IsExpr exValue (Maybe value), IsExpr exBigMap (BigMap key value), IsComparable key, TypeValue (BigMap key value)) => exKey -> exValue -> exBigMap -> Expr (BigMap key value)

EmptyBigMap :: (KnownCValue key, KnownValue value, TypeValue (BigMap key value)) => Expr (BigMap key value)

Pack :: (IsExpr ex a, NicePackedValue a) => ex -> Expr ByteString

Unpack :: (NiceUnpackedValue a, Typeable a, IsExpr bsExpr ByteString) => bsExpr -> Expr (Maybe a)

Cons :: (IsExpr ex1 a, IsExpr ex2 (List a), TypeValue a) => ex1 -> ex2 -> Expr (List a)

Nil :: (KnownValue a, TypeValue a) => Expr (List a)

Concat :: (IsExpr ex c, ConcatOpHs c, TypeValue c) => ex -> ex -> Expr c

Concat' :: (IsExpr ex (List c), ConcatOpHs c, TypeValue c) => ex -> Expr c

Slice :: (IsExpr ex c, SliceOpHs c, TypeValue c, IsExpr an Natural, IsExpr bn Natural) => an -> bn -> ex -> Expr (Maybe c)

Contract :: (NiceParameterFull p, NoExplicitDefaultEntryPoint p, ToTAddress p addr, ToT addr ~ ToT Address, IsExpr exAddr addr) => exAddr -> Expr (Maybe (ContractRef p))

Self :: (NiceParameterFull p, NoExplicitDefaultEntryPoint p) => Expr (ContractRef p)

ContractAddress :: IsExpr exc (ContractRef p) => exc -> Expr Address

ContractCallingUnsafe :: (NiceParameter arg, Typeable arg, IsExpr exAddr Address) => EpName -> exAddr -> Expr (Maybe (ContractRef arg))

RunFutureContract :: (NiceParameter p, Typeable p, IsExpr conExpr (FutureContract p)) => conExpr -> Expr (Maybe (ContractRef p))

ImplicitAccount :: IsExpr exkh KeyHash => exkh -> Expr (ContractRef ())

ConvertEpAddressToContract :: (NiceParameter p, Typeable p, IsExpr epExpr EpAddress) => epExpr -> Expr (Maybe (ContractRef p))

CheckSignature :: (IsExpr pkExpr PublicKey, IsExpr sigExpr Signature, IsExpr hashExpr ByteString) => pkExpr -> sigExpr -> hashExpr -> Expr Bool

Sha256 :: IsExpr hashExpr ByteString => hashExpr -> Expr ByteString

Sha512 :: IsExpr hashExpr ByteString => hashExpr -> Expr ByteString

Blake2b :: IsExpr hashExpr ByteString => hashExpr -> Expr ByteString

HashKey :: IsExpr keyExpr PublicKey => keyExpr -> Expr KeyHash

ChainId :: Expr ChainId

Now :: Expr Timestamp

Amount :: Expr Mutez

Balance :: Expr Mutez

Sender :: Expr Address

Generalization of Expr

type IsExpr op n = (ToExpr op, ExprType op ~ n)

type AreExprs ex1 ex2 n m = (IsExpr ex1 n, IsExpr ex2 m)

class ToExpr' (Decide x) x => ToExpr x

type ExprType a = ExprType' (Decide a) a

type (:~>) op n = IsExpr op n

toExpr :: forall a. ToExpr a => a -> Expr (ExprType a)

Expr compilation

compileExpr :: forall a inp. Expr a -> IndigoM inp (a & inp) ()

compileToExpr :: ToExpr a => a -> IndigoM inp (ExprType a & inp) ()

nullaryOp :: TypeValue res => (inp :-> (res ': inp)) -> IndigoM inp (res ': inp) ()

unaryOp :: forall n ex res inp. (IsExpr ex n, TypeValue res) => ex -> ((n & inp) :-> (res & inp)) -> IndigoM inp (res & inp) ()

binaryOp :: forall n m ex1 ex2 res inp. (AreExprs ex1 ex2 n m, TypeValue res) => ex1 -> ex2 -> ((n & (m & inp)) :-> (res & inp)) -> IndigoM inp (res & inp) ()

ternaryOp :: forall n m l ex1 ex2 ex3 res inp. (AreExprs ex1 ex2 n m, IsExpr ex3 l, TypeValue res) => ex1 -> ex2 -> ex3 -> ((n & (m & (l & inp))) :-> (res & inp)) -> IndigoM inp (res & inp) ()

nullaryOpFlat :: (inp :-> inp) -> IndigoM inp inp ()

unaryOpFlat :: forall n ex inp. IsExpr ex n => ex -> ((n & inp) :-> inp) -> IndigoM inp inp ()

binaryOpFlat :: forall n m ex1 ex2 inp. AreExprs ex1 ex2 n m => ex1 -> ex2 -> ((n & (m & inp)) :-> inp) -> IndigoM inp inp ()

ternaryOpFlat :: forall n m l ex1 ex2 ex3 inp. (AreExprs ex1 ex2 n m, IsExpr ex3 l) => ex1 -> ex2 -> ex3 -> ((n & (m & (l & inp))) :-> inp) -> IndigoM inp inp ()

Basic

constExpr :: (Typeable a, NiceConstant a) => a -> Expr a

Create an expression holding a constant.

varExpr :: TypeValue a => Var a -> Expr a

Create an expression holding a variable.

cast :: (ex :~> a, KnownValue a, Typeable a) => ex -> Expr a

mem :: (MemOpHs c, IsExpr exc c, IsExpr exck (MemOpKeyHs c)) => exck -> exc -> Expr Bool

Math

(+.) :: (AreExprs ex1 ex2 n m, ArithOpHs Add n m, TypeValue (ArithResHs Add n m)) => ex1 -> ex2 -> Expr (ArithResHs Add n m) infixl 6

(-.) :: (AreExprs ex1 ex2 n m, ArithOpHs Sub n m, TypeValue (ArithResHs Sub n m)) => ex1 -> ex2 -> Expr (ArithResHs Sub n m) infixl 6

(*.) :: (AreExprs ex1 ex2 n m, ArithOpHs Mul n m, TypeValue (ArithResHs Mul n m)) => ex1 -> ex2 -> Expr (ArithResHs Mul n m) infixl 7

(/.) :: (AreExprs ex1 ex2 n m, EDivOpHs n m, TypeValue (EDivOpResHs n m)) => ex1 -> ex2 -> Expr (EDivOpResHs n m) infixl 7

(%.) :: (AreExprs ex1 ex2 n m, EDivOpHs n m, TypeValue (EModOpResHs n m)) => ex1 -> ex2 -> Expr (EModOpResHs n m) infixl 7

neg :: (ex :~> n, UnaryArithOpHs Neg n, TypeValue (UnaryArithResHs Neg n)) => ex -> Expr (UnaryArithResHs Neg n)

abs :: (ex :~> n, UnaryArithOpHs Abs n, TypeValue (UnaryArithResHs Abs n)) => ex -> Expr (UnaryArithResHs Abs n)

Comparison

(==.) :: (NiceComparable n, AreExprs c c1 n n) => c -> c1 -> Expr Bool infix 4

(/=.) :: (NiceComparable n, AreExprs c c1 n n) => c -> c1 -> Expr Bool infix 4

(<=.) :: (NiceComparable n, AreExprs c c1 n n) => c -> c1 -> Expr Bool infix 4

(>=.) :: (NiceComparable n, AreExprs c c1 n n) => c -> c1 -> Expr Bool infix 4

(<.) :: (NiceComparable n, AreExprs c c1 n n) => c -> c1 -> Expr Bool infix 4

(>.) :: (NiceComparable n, AreExprs c c1 n n) => c -> c1 -> Expr Bool infix 4

Conversion

isNat :: ex :~> Integer => ex -> Expr (Maybe Natural)

int :: ex :~> Natural => ex -> Expr Integer

Bits and boolean

(<<.) :: (AreExprs ex1 ex2 n m, ArithOpHs Lsl n m, TypeValue (ArithResHs Lsl n m)) => ex1 -> ex2 -> Expr (ArithResHs Lsl n m) infixl 8

(>>.) :: (AreExprs ex1 ex2 n m, ArithOpHs Lsr n m, TypeValue (ArithResHs Lsr n m)) => ex1 -> ex2 -> Expr (ArithResHs Lsr n m) infixl 8

(&&.) :: (AreExprs ex1 ex2 n m, ArithOpHs And n m, TypeValue (ArithResHs And n m)) => ex1 -> ex2 -> Expr (ArithResHs And n m) infixr 3

(||.) :: (AreExprs ex1 ex2 n m, ArithOpHs Or n m, TypeValue (ArithResHs Or n m)) => ex1 -> ex2 -> Expr (ArithResHs Or n m) infixr 2

not :: (op :~> n, UnaryArithOpHs Not n, TypeValue (UnaryArithResHs Not n)) => op -> Expr (UnaryArithResHs Not n)

xor :: (AreExprs ex1 ex2 n m, ArithOpHs Xor n m, TypeValue (ArithResHs Xor n m)) => ex1 -> ex2 -> Expr (ArithResHs Xor n m)

Serialization

pack :: (IsExpr ex a, NicePackedValue a) => ex -> Expr ByteString

unpack :: (NiceUnpackedValue a, Typeable a, bsExpr :~> ByteString) => bsExpr -> Expr (Maybe a)

Pairs

pair :: (AreExprs ex1 ex2 n m, TypeValue (n, m)) => ex1 -> ex2 -> Expr (n, m)

fst :: (op :~> (n, m), TypeValue n) => op -> Expr n

snd :: (op :~> (n, m), TypeValue m) => op -> Expr m

Maybe

some :: (ex :~> t, TypeValue (Maybe t)) => ex -> Expr (Maybe t)

none :: (TypeValue t, KnownValue t) => Expr (Maybe t)

Bytes, String

slice :: (ex :~> c, an :~> Natural, bn :~> Natural, SliceOpHs c, TypeValue c) => (an, bn) -> ex -> Expr (Maybe c)

concat :: (IsExpr ex c, ConcatOpHs c, TypeValue c) => ex -> ex -> Expr c

(<>.) :: (IsExpr ex c, ConcatOpHs c, TypeValue c) => ex -> ex -> Expr c infixr 6

List

(.:) :: (ex1 :~> a, ex2 :~> List a, TypeValue a) => ex1 -> ex2 -> Expr (List a) infixr 5

concat' :: (IsExpr ex (List c), ConcatOpHs c, TypeValue c) => ex -> Expr c

nil :: (KnownValue a, TypeValue a) => Expr (List a)

Store

(%@) :: (HasUStore name key value store, TypeValue value, exKey :~> key, exStore :~> UStore store) => exStore -> (Label name, exKey) -> Expr (Maybe value) infixr 8

(%~~) :: (HasUStore name key value store, IsError err, exKey :~> key, exVal :~> value, exStore :~> UStore store, TypeValue (UStore store)) => exStore -> (Label name, err, exKey, exVal) -> Expr (UStore store) infixr 8

(%~) :: (HasUStore name key value store, TypeValue (UStore store), exKey :~> key, exVal :~> value, exStore :~> UStore store) => exStore -> (Label name, exKey, exVal) -> Expr (UStore store) infixr 8

(%!) :: (StoreHasField store fname ftype, exStore :~> store, TypeValue ftype) => exStore -> Label fname -> Expr ftype infixr 8

Record

(~.) :: (InstrSetFieldC dt name, IsExpr exDt dt, exFld :~> GetFieldType dt name, TypeValue dt) => exDt -> (Label name, exFld) -> Expr dt infixl 8

(!.) :: (InstrGetFieldC dt name, exDt :~> dt, TypeValue (GetFieldType dt name)) => exDt -> Label name -> Expr (GetFieldType dt name) infixl 8

(!~) :: (ex :~> t, TypeValue (name :! t)) => ex -> Label name -> Expr (name :! t)

construct :: (InstrConstructC dt, RMap (ConstructorFieldTypes dt), TypeValue dt) => Rec Expr (ConstructorFieldTypes dt) -> Expr dt

constructT :: (InstrConstructC dt, TypeValue dt, RMap (ConstructorFieldTypes dt), fields ~ Rec Expr (ConstructorFieldTypes dt), RecFromTuple fields) => IsoRecTuple fields -> Expr dt

Map

(^?) :: (GetOpHs c, TypeValue (Maybe (GetOpValHs c)), exKey :~> GetOpKeyHs c, exMap :~> c) => exMap -> exKey -> Expr (Maybe (GetOpValHs c)) infixl 8

(^~) :: (UpdOpHs c, TypeValue c, exKey :~> UpdOpKeyHs c, exVal :~> UpdOpParamsHs c, exMap :~> c) => exMap -> (exKey, exVal) -> Expr c infixl 8

msize :: (SizeOpHs c, exMap :~> c) => exMap -> Expr Natural

BigMap

(&?) :: (exKey :~> key, exBigMap :~> BigMap key value, IsComparable key, TypeValue (Maybe value)) => exBigMap -> exKey -> Expr (Maybe value) infixl 8

(&~) :: (exKey :~> key, exValue :~> value, exBigMap :~> BigMap key value, IsComparable key, TypeValue value, TypeValue (BigMap key value)) => exBigMap -> (exKey, exValue) -> Expr (BigMap key value) infixl 8

(&?~) :: (exKey :~> key, exValue :~> Maybe value, exBigMap :~> BigMap key value, IsComparable key, TypeValue (BigMap key value)) => exBigMap -> (exKey, exValue) -> Expr (BigMap key value) infixl 8

(&-) :: (TypeValue value, KnownValue value, exKey :~> key, exBigMap :~> BigMap key value, IsComparable key, TypeValue (BigMap key value)) => exBigMap -> exKey -> Expr (BigMap key value) infixl 8

emptyBigMap :: (KnownCValue key, KnownValue value, TypeValue (BigMap key value)) => Expr (BigMap key value)

Set

(#?) :: (exKey :~> key, exSet :~> Set key, IsComparable key) => exSet -> exKey -> Expr Bool infixr 8

(#~) :: (exKey :~> key, exSet :~> Set key, IsComparable key, TypeValue (Set key)) => exSet -> exKey -> Expr (Set key) infixr 8

(#-) :: (exKey :~> key, exSet :~> Set key, IsComparable key, TypeValue (Set key)) => exSet -> exKey -> Expr (Set key) infixr 8

ssize :: exSet :~> Set key => exSet -> Expr Natural

Contract

contract :: (NiceParameterFull p, NoExplicitDefaultEntryPoint p, ToTAddress p addr, ToT addr ~ ToT Address, exAddr :~> addr) => exAddr -> Expr (Maybe (ContractRef p))

self :: (NiceParameterFull p, NoExplicitDefaultEntryPoint p) => Expr (ContractRef p)

contractAddress :: exc :~> ContractRef p => exc -> Expr Address

contractCallingUnsafe :: (NiceParameter arg, Typeable arg, exAddr :~> Address) => EpName -> exAddr -> Expr (Maybe (ContractRef arg))

runFutureContract :: (NiceParameter p, Typeable p, conExpr :~> FutureContract p) => conExpr -> Expr (Maybe (ContractRef p))

implicitAccount :: exkh :~> KeyHash => exkh -> Expr (ContractRef ())

convertEpAddressToContract :: (NiceParameter p, Typeable p, epExpr :~> EpAddress) => epExpr -> Expr (Maybe (ContractRef p))

Auxiliary

now :: Expr Timestamp

amount :: Expr Mutez

sender :: Expr Address

blake2b :: hashExpr :~> ByteString => hashExpr -> Expr ByteString

sha256 :: hashExpr :~> ByteString => hashExpr -> Expr ByteString

sha512 :: hashExpr :~> ByteString => hashExpr -> Expr ByteString

hashKey :: keyExpr :~> PublicKey => keyExpr -> Expr KeyHash

chainId :: Expr ChainId

balance :: Expr Mutez

checkSignature :: (pkExpr :~> PublicKey, sigExpr :~> Signature, hashExpr :~> ByteString) => pkExpr -> sigExpr -> hashExpr -> Expr Bool

To make extensions work

fromLabel :: IsLabel x a => a