Indigo.Language

Description

High level statements of Indigo language.

Documentation

module Indigo.Language.Scope

module Indigo.Language.Conditional

module Indigo.Language.Error

module Indigo.Language.Case

Vars

newVar :: (IsExpr ex x, TypeValue x) => ex -> IndigoM inp (x & inp) (Var x)

Create a new variable with passed expression as an initial value.

setVar :: (IsExpr ex x, TypeValue x) => Var x -> ex -> IndigoM inp inp ()

Set to variable new value.

updateVar :: (IsExpr ey y, TypeValue x, TypeValue y) => ([y, x] :-> '[x]) -> Var x -> ey -> IndigoM inp inp ()

Call binary operator with constant argument to update variable in-place.

setField_ :: forall dt name inp ex. (InstrSetFieldC dt name, TypeValue dt, IsExpr ex (GetFieldType dt name), TypeValue (GetFieldType dt name)) => Var dt -> Label name -> ex -> IndigoM inp inp ()

Loop

forEach :: forall a e inp xs. (IterOpHs a, TypeValue (IterOpElHs a), e :~> a) => e -> (Var (IterOpElHs a) -> IndigoM (IterOpElHs a & inp) xs ()) -> IndigoM inp inp ()

For statements to iterate over container.

while :: forall inp xs. Expr Bool -> IndigoM inp xs () -> IndigoM inp inp ()

While statement. The same rule about releasing.

Procedural

(+=) :: (IsExpr ex1 n, ArithOpHs Add n m, ArithResHs Add n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(-=) :: (IsExpr ex1 n, ArithOpHs Sub n m, ArithResHs Sub n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(*=) :: (IsExpr ex1 n, ArithOpHs Mul n m, ArithResHs Mul n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(<<=) :: (IsExpr ex1 n, ArithOpHs Lsl n m, ArithResHs Lsl n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(>>=.) :: (IsExpr ex1 n, ArithOpHs Lsr n m, ArithResHs Lsr n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(&&=) :: (IsExpr ex1 n, ArithOpHs And n m, ArithResHs And n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(||=) :: (IsExpr ex1 n, ArithOpHs Or n m, ArithResHs Or n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

(^=) :: (IsExpr ex1 n, ArithOpHs Xor n m, ArithResHs Xor n m ~ m, TypeValue m, TypeValue n) => Var m -> ex1 -> IndigoM inp inp ()

Contract call

selfCalling :: forall inp p mname. (NiceParameterFull p, TypeValue (GetEntryPointArgCustom p mname)) => EntryPointRef mname -> IndigoM inp (ContractRef (GetEntryPointArgCustom p mname) & inp) ()

contractCalling :: forall inp cp epRef epArg addr exAddr. (HasEntryPointArg cp epRef epArg, ToTAddress cp addr, ToT addr ~ ToT Address, IsExpr exAddr addr, TypeValue epArg) => epRef -> exAddr -> IndigoM inp (Maybe (ContractRef epArg) & inp) ()

Documentation

doc :: DocItem di => di -> IndigoM s s ()

Put a document item.

docGroup :: DocGrouping -> IndigoM i o () -> IndigoM i o ()

Group documentation built in the given piece of code into block dedicated to one thing, e.g. to one entrypoint.

contractName :: Text -> IndigoM i o () -> IndigoM i o ()

Give a name to given contract. Apply it to the whole contract code.

Entrypoints

type IndigoContractCode param store = (HasStorage store, HasSideEffects) => Var param -> IndigoProcedure '[param, store, Ops]

type IndigoEntrypoint param = forall s. param -> IndigoProcedure s

transferTokens :: (IsExpr exp p, IsExpr exm Mutez, IsExpr exc (ContractRef p), NiceParameter p, HasSideEffects) => exp -> exm -> exc -> IndigoM inp inp ()

setDelegate :: (HasSideEffects, IsExpr ex (Maybe KeyHash)) => ex -> IndigoM inp inp ()

createContract :: (TypeValue st, IsExpr exk (Maybe KeyHash), IsExpr exm Mutez, IsExpr exs st, NiceStorage st, NiceParameterFull param, HasSideEffects) => (HasStorage st => Var param -> IndigoM '[param, st, Ops] out ()) -> exk -> exm -> exs -> IndigoM inp (Address & inp) (Var Address)

Functions and Procedures

type IndigoFunction s ret = IndigoM s (ScopeOutStack ret s) (ScopeRetVars ret)

Utility type for an IndigoM that add one element to the stack and returns a variable pointing at it.

type IndigoProcedure s = IndigoM s s ()

Utility type for an IndigoM that does not modify the stack (only the values in it) and returns nothing.

scope :: forall a inp out. ScopeCodeGen a => IndigoM inp out a -> IndigoFunction inp a

Takes an arbitrary IndigoM and wraps it into an IndigoFunction producing a local scope for its execution. Once it executed, all non-returned variables are cleaned up so that the stack has only returned variables at the top. This also can be interpreted as `if True then f else nop`.

Note, that by default we do not define scope inside indigo functions, meaning that once we want to create a new variable or return it from a function we need to do it inside `scope $ instr` construction, for example:

``` f :: IndigoFunction s Natural f = scope $ do *[s]* res <- newVar (0 :: Natural) *[Natural, s]* scope $ do _n <- newVar (1 :: Integer) *[Integer, Natural, s] res += 4 *[Natural, s]* return res *[s]* ```

defFunction :: forall a inp out. ScopeCodeGen a => IndigoM inp out a -> IndigoFunction inp a

Alias for scope we use in a tutorial.

defContract :: ((HasStorage st, HasSideEffects) => IndigoM '[param, st, Ops] out ()) -> (HasStorage st, HasSideEffects) => IndigoProcedure '[param, st, Ops]

A more specific version of defFunction meant to more easily create IndigoContracts.

Conversion from Lorentz

fromLorentzFun1 :: (TypeValue ret, IsExpr ex a) => (:->) ((&) a s) ((&) ret s) -> ex -> IndigoM s ((&) ret s) (Var ret)

fromLorentzFun2 :: (TypeValue ret, IsExpr ex a, IsExpr ex a) => (:->) ((&) a ((&) a s)) ((&) ret s) -> ex -> ex -> IndigoM s ((&) ret s) (Var ret)

fromLorentzFun3 :: (TypeValue ret, IsExpr ex a, IsExpr ex a, IsExpr ex a) => (:->) ((&) a ((&) a ((&) a s))) ((&) ret s) -> ex -> ex -> ex -> IndigoM s ((&) ret s) (Var ret)

fromLorentzFun1Void :: IsExpr ex a => (:->) ((&) a s) s -> ex -> IndigoM s s ()

fromLorentzFun2Void :: (IsExpr ex a, IsExpr ex a) => (:->) ((&) a ((&) a s)) s -> ex -> ex -> IndigoM s s ()

fromLorentzFun3Void :: (IsExpr ex a, IsExpr ex a, IsExpr ex a) => (:->) ((&) a ((&) a ((&) a s))) s -> ex -> ex -> ex -> IndigoM s s ()