Indigo.Language.Case

Description

High level statements of Indigo language.

Documentation

(#=) :: CaseArrow name body clause => Label name -> body -> clause

An alias for /-> to use when the body is a lambda so that the whole thing won't have multiple arrows.

(//->) :: (CaseArrow name (Var x -> IndigoAnyOut x inp ret) (IndigoCaseClauseL inp ret (CaseClauseParam ctor field)), ScopeCodeGen retBr, ret ~ ScopeRetExprs retBr, ScopeOutStack ret inp ~ ScopeOutStack retBr inp) => Label name -> (Var x -> IndigoM (x ': inp) xs retBr) -> IndigoCaseClauseL inp ret (CaseClauseParam ctor field) infixr 0

Use this instead of /->. This operator is like /-> but wrap a body into IndigoAnyOut, which is needed for two reasons: to allow having any output stack and to allow to return not exactly the same values.

case_ :: forall dt guard inp ret clauses. CaseCommon dt guard inp ret clauses => guard -> clauses -> IndigoM inp (ScopeOutStack ret inp) (ScopeRetVars ret)

A case statement for indigo. See examples for a sample usage.

caseT :: forall dt guard inp ret clauses. (CaseCommon dt guard inp ret clauses, RecFromTuple clauses) => guard -> IsoRecTuple clauses -> IndigoM inp (ScopeOutStack ret inp) (ScopeRetVars ret)

case_ for tuples.

entryCase :: forall dt entryPointKind guard inp ret clauses. (CaseCommon dt guard inp ret clauses, RecFromTuple clauses, DocumentEntryPoints entryPointKind dt) => Proxy entryPointKind -> guard -> IsoRecTuple clauses -> IndigoM inp (ScopeOutStack ret inp) (ScopeRetVars ret)

entryCase_ for tuples.

entryCase_ :: forall dt entryPointKind guard inp ret clauses. (CaseCommon dt guard inp ret clauses, DocumentEntryPoints entryPointKind dt) => Proxy entryPointKind -> guard -> clauses -> IndigoM inp (ScopeOutStack ret inp) (ScopeRetVars ret)

case_ for pattern-matching on parameter.

entryCaseSimple :: forall cp guard inp ret clauses. (CaseCommon cp guard inp ret clauses, RecFromTuple clauses, DocumentEntryPoints PlainEntryPointsKind cp, NiceParameterFull cp, RequireFlatParamEps cp) => guard -> IsoRecTuple clauses -> IndigoM inp (ScopeOutStack ret inp) (ScopeRetVars ret)

entryCase for contracts with flat parameter.