Indigo.Language.Conditional

Description

Conditional statements of Indigo language.

Documentation

if_ :: forall inp xs ys a b exc. (IfConstraint a b inp, exc :~> Bool) => exc -> IndigoM inp xs a -> IndigoM inp ys b -> IndigoM inp (ScopeOutStack a inp) (ScopeRetVars a)

If statement. It's supposed that all created variables inside branches will be released after execution leaves a scope where they were created.

ifJust :: forall inp xs ys x a b exa. (IfConstraint a b inp, TypeValue x, exa :~> Maybe x) => exa -> (Var x -> IndigoM (x & inp) xs a) -> IndigoM inp ys b -> IndigoM inp (ScopeOutStack a inp) (ScopeRetVars a)

If which works like case for Maybe.

when :: exc :~> Bool => exc -> IndigoM inp out () -> IndigoM inp inp ()

Run the instruction when condition is met, do nothing otherwise.

whenJust :: forall inp xs x exa. (TypeValue x, exa :~> Maybe x) => exa -> (Var x -> IndigoM (x & inp) xs ()) -> IndigoM inp inp ()

Run the instruction when the given expression returns Just value or do nothing otherwise.

whenNothing :: forall inp x ys exa. (TypeValue x, exa :~> Maybe x) => exa -> IndigoM inp ys () -> IndigoM inp inp ()

Run the instruction when the given expression returns Nothing or do nothing otherwise.

type IfConstraint a b inp = (ScopeCodeGen a, ScopeCodeGen b, CompareBranchesResults (ScopeRetExprs a) (ScopeRetExprs b), ScopeOutStack a inp ~ ScopeOutStack b inp)