Indigo.Language.Scope

Description

Machinery to provide ability to return values from Indigo statements.

Documentation

type ScopeCodeGen ret = ScopeCodeGen' (ClassifyBranchReturn ret) ret

class ScopeCodeGen' (retKind :: BranchRetKind) (ret :: Type) where

Type class which unions all related management of computations in a scope, like in 'if' branch, in 'case' body, etc. Particularly, it takes care of the computation of expressions returning from a scope to leave the scope safely. Basically, this type class encapsulates generation of Lorentz code which looks like: branch_code # we get some arbitrary type of a stack here, lets call it xs compute_returning_expressions # we get type of stack [e1, e2, ... ek] ++ xs cleanup_xs_to_inp we get [e1, e2, e3, ..., ek] ++ inp

Associated Types

type ScopeOutStack' retKind ret (inp :: [Type]) :: [Type]

Type family reflecting the output stack of the resulting statement.

type ScopeRetVars' retKind ret :: Type

Type family reflecting the output value of the resulting statement.

type ScopeRetExprs' retKind ret :: Type

Tuple looking like `(Expr x, Expr y, ..)`, which corresponds to expressions returning from the scope. ScopeRetVars' and ScopeRetExprs' are twin types because the former just adds Var over each expression from the later one.

Methods

compileScopeReturn' :: ret -> IndigoM xs (ScopeOutStack' retKind ret xs) ()

Produces an Indigo computation, which puts on a stack evaluated returnen expressions from the leaving scope.

liftClear' :: (xs :-> inp) -> ScopeOutStack' retKind ret xs :-> ScopeOutStack' retKind ret inp

Drop unused cells of a stack from the leaving scope, which are after evaluated expressions.

finalizeStatement' :: MetaData inp -> (inp :-> ScopeOutStack' retKind ret inp) -> GenCode inp (ScopeOutStack' retKind ret inp) (ScopeRetVars' retKind ret)

Push a variables in MetaData, which refer to generated expressions, and generate gcClear for the whole statement

type ScopeOutStack ret inp = ScopeOutStack' (ClassifyBranchReturn ret) ret inp

type ScopeRetVars ret = ScopeRetVars' (ClassifyBranchReturn ret) ret

type ScopeRetExprs ret = ScopeRetExprs' (ClassifyBranchReturn ret) ret

type family ClassifyBranchReturn (ret :: Type) where ...

This type family returns promoted value of type BranchRetKind or cause a compilation errors, if tuple with too many elements is used.

Equations

ClassifyBranchReturn () = Unit

ClassifyBranchReturn (_, _) = Tuple

ClassifyBranchReturn (_, _, _) = Tuple

ClassifyBranchReturn (_, _, _, _) = TypeError (Text "Tuple with 4 elements is not supported yet as returning value")

ClassifyBranchReturn (_, _, _, _, _) = TypeError (Text "Tuple with 5 elements is not supported yet as returning value")

ClassifyBranchReturn (_, _, _, _, _, _) = TypeError (Text "Tuple with 6 elements is not supported yet as returning value")

ClassifyBranchReturn _ = SingleVal

compileScope :: forall ret inp xs. ScopeCodeGen ret => GenCode inp xs ret -> inp :-> ScopeOutStack ret inp

Concatenate a scope code, generation of returning expressions, and clean up of redundant cells from a stack

finalizeStatement :: forall ret inp. ScopeCodeGen ret => MetaData inp -> (inp :-> ScopeOutStack ret inp) -> GenCode inp (ScopeOutStack ret inp) (ScopeRetVars ret)

Specify type of finalizeStatement'