Indigo.Lib

Description

Standard library for Indigo.

As opposed to Expr and Language modules, this module contains quite standard things that are not present in vanilla Michelson and are not typically found in imperative high level languages.

Documentation

view_ :: forall arg r viewExpr inp out. (TypeValue arg, IsoValue arg, NiceParameter r, Typeable r, viewExpr :~> View arg r, HasSideEffects) => (Expr arg -> IndigoM inp out (Expr r)) -> viewExpr -> IndigoM inp out ()

Indigo version of the view macro. It takes a function from view argument to view result and View structure that typically comes from top-level case.

project :: forall arg r viewExpr inp out. (TypeValue arg, IsoValue arg, NiceParameter r, Typeable r, viewExpr :~> View arg r, HasSideEffects) => viewExpr -> (Expr arg -> IndigoM inp out (Expr r)) -> IndigoM inp out ()

Flipped version of view_ that is present due to the common appearance of `flip view parameter $ instr` construction. Semantically we "project" the given parameter inside the body of View construction.

subGt0 :: (ex1 :~> Natural, ex2 :~> Natural) => ex1 -> ex2 -> (forall st. IndigoM st st ()) -> IndigoFunction s (Maybe Natural)