Indigo.Lookup

Description

This module contains the logic to lookup Vars in a stack and the actions to manipulate it.

For efficiency, actions are implemented using Lorentz macros. To do so every necessary constraint is checked at runtime.

Variable Lookup Actions

varActionGet :: TypeValue a => Var a -> StackVars stk -> stk :-> (a & stk)

Puts a copy of the value for the given Var on top of the stack

varActionSet :: TypeValue a => Var a -> StackVars stk -> (a & stk) :-> stk

Sets the value for the given Var to the topmost value on the stack

varActionUpdate :: (TypeValue a, TypeValue b) => Var a -> StackVars stk -> ('[b, a] :-> '[a]) -> (b ': stk) :-> stk

Updates the value for the given Var with the topmost value on the stack using the given binary instruction.

varActionOperation :: HasSideEffects => StackVars stk -> (Operation ': stk) :-> stk

Given a stack with a list of Operations on its bottom, updates it by appending the Operation on the top.

Vinyl manipulation helpers

rtake :: Sing n -> Rec any s -> Rec any (Take n s)

rdrop :: Sing n -> Rec any s -> Rec any (Drop n s)