Indigo.Print

Description

Module containing pretty-printing of Indigo contracts

Documentation

printIndigoContract

Arguments

:: (TypeValue param, TypeValue st, NiceParameterFull param, NiceStorage st)

=> Bool
Force result to be single line

-> IndigoContract' param st out

-> LText

Pretty-print an Indigo contract into Michelson code.

printAsMichelson :: (TypeValue param, TypeValue st, NiceParameterFull param, NiceStorage st, MonadIO m) => IndigoContract' param st out -> m ()

Prints the pretty-printed Michelson code of an Indigo contract to the standard output. This is intended to be easy to use for newcomers.

saveAsMichelson :: (TypeValue param, TypeValue st, NiceParameterFull param, NiceStorage st, MonadIO m, MonadMask m) => IndigoContract' param st out -> FilePath -> m ()

Saves the pretty-printed Michelson code of an Indigo contract to the given file. This is intended to be easy to use for newcomers.