Indigo.Rebinded

Description

Reimplementation of some syntax sugar

You need the following module pragmas to make it work smoothly:

{--} {--}

Documentation

ifThenElse :: (IfConstraint a b inp, IsExpr exa Bool) => exa -> IndigoM inp xs a -> IndigoM inp ys b -> IndigoM inp (ScopeOutStack a inp) (ScopeRetVars a)

Defines semantics of if ... then ... else ... construction for Indigo where the predicate is a generic exa for which `IsExpr exa Bool` holds

class IsLabel (x :: Symbol) a where

Methods

fromLabel :: a