Indigo.State

Description

This module contains core of Indigo language: IndigoM monad, datatype to represent its state and variables. It also includes some convenient functions to work with state in IndigoM, to provide rebindable syntax.

IndigoM monad implements functionality of a symbolic interpreter. During its execution Lorentz code is being generated.

Documentation

newtype IndigoM inp out a

IndigoM monad. It's basically https://hackage.haskell.org/package/category-extras-0.53.5/docs/Control-Monad-Indexed-State.html however, this package is not in the used lts and it doesn't compile.

Constructors

IndigoM

Fields

runIndigoM :: MetaData inp -> GenCode inp out a

(>>=) :: IndigoM inp out a -> (a -> IndigoM out out1 b) -> IndigoM inp out1 b

Bind for rebindable syntax. It's basically like bind for State monad but also compose generated code from _m a_ and _a -> m b_.

(=<<) :: (a -> IndigoM out out1 b) -> IndigoM inp out a -> IndigoM inp out1 b

(>>) :: IndigoM inp out a -> IndigoM out out1 b -> IndigoM inp out1 b

Then for rebindable syntax.

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

An infix synonym for fmap.

The name of this operator is an allusion to $. Note the similarities between their types:

 ($)  ::              (a -> b) ->   a ->   b
(<$>) :: Functor f => (a -> b) -> f a -> f b

Whereas $ is function application, <$> is function application lifted over a Functor.

Examples

Expand

Convert from a Maybe Int to a Maybe String using show:

>>> show <$> Nothing
Nothing
>>> show <$> Just 3
Just "3"

Convert from an Either Int Int to an Either Int String using show:

>>> show <$> Left 17
Left 17
>>> show <$> Right 17
Right "17"

Double each element of a list:

>>> (*2) <$> [1,2,3]
[2,4,6]

Apply even to the second element of a pair:

>>> even <$> (2,2)
(2,True)

return :: a -> IndigoM inp inp a

Return for rebindable syntax.

iget :: IndigoM inp inp (MetaData inp)

Get current MetaData.

iput :: GenCode inp out a -> IndigoM inp out a

Put new GenCode.

type TypeValue a = (Typeable a, Typeable (ToT a))

Constraint for StkEl, required for comparison in constraint checking

newtype Var a

Variable whcih is exposed to a user. It stores number of variable, its phantom type represent type of a variable.

Constructors

Var Word

data StkEl a

Stack element of symbolic interpreter. It holds either a variable which refers to this element or just NoRef which indicates that there are no references to this element.

Constructors

TypeValue a => NoRef

TypeValue a => Ref (Var a)

type StackVars (stk :: [Type]) = Rec StkEl stk

Stack of symbolic interpreter.

data GenCode inp out a

Resulting state of IndigoM.

Constructors

GenCode

Fields

gcOut :: ~a

Interpreter output value

gcMeta :: ~(MetaData out)

Interpreter meta data.

gcCode :: inp :-> out

Generated Lorentz code.

gcClear :: out :-> inp

Clearing Lorentz code.

data MetaData stk

Initial state of IndigoM.

Constructors

MetaData

Fields

mdStack :: StackVars stk

Stack of symbolic interpreter.

mdRefCount :: Word

Number of allocated variables.

makeTopVar :: TypeValue x => IndigoM (x & inp) (x & inp) (Var x)

Create a variable which references to the top of a stack.

pushRefMd :: TypeValue x => MetaData stk -> (Var x, MetaData (x & stk))

Push a new stack element with reference to it. Return a variable which references to the just added element.

pushNoRefMd :: TypeValue a => MetaData inp -> MetaData (a & inp)

Push a new stack element without reference to it.

popNoRefMd :: MetaData (a & inp) -> MetaData inp

Remove top element of the stack. It's supposed that no variables refer to the top of the stack.

withVarAt :: (TypeValue a, a ~ At n inp, RequireLongerThan inp n) => MetaData inp -> Sing n -> (MetaData inp, Var a)

Given a MetaData and a Peano singleton for a depth, it puts a new Var at that depth (0-indexed) and returns it with the updated MetaData. If there is a Var there already it is used and the MetaData not changed.

cleanGenCode :: GenCode inp out a -> inp :-> inp

Produces the generated Lorentz code that cleans after itself, leaving the same stack as the input one

type DefaultStack stk = Default (MetaData stk)